[SERVICE-507] Move contextChannel dependencies out of connection
