+++ test/demo/utils/fdc3RemoteExecution.ts
@@ -13,9 +13,8 @@
 import {Identity} from 'openfin/_v2/main';
 import {WindowOption} from 'openfin/_v2/api/window/windowOption';
 
-import {Application, Context, IntentType, AppIntent, ChannelId, DefaultChannel, DesktopChannel, Channel} from '../../../src/client/main';
+import {Application, Context, IntentType, AppIntent, ChannelId, DefaultChannel, DesktopChannel, Channel, FDC3Event, FDC3EventType} from '../../../src/client/main';
 import {RaiseIntentPayload} from '../../../src/client/internal';
-import {FDC3Event, FDC3EventType} from '../../../src/client/connection';
 
 import {OFPuppeteerBrowser, TestWindowContext, TestChannelTransport} from './ofPuppeteer';
